save scraped data with userID
