+++ app/public/analyse.html
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <title>Digital Mirror</title>
+    <link href="styles/style.css" rel="stylesheet">
+  </head>
+  <body>
+  <div id="background">
+    <div id='home'>
+      <div class="typeText">
+        <h1>Analysing Your Data...</h1>
+      </div>
+      <p>doo doo doo</p>
+    </div>
+  </div>
+  </body>
+</html>
